% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chung_2021.R
\name{Chung_2021}
\alias{Chung_2021}
\title{Obtain the Chung_2021 data}
\usage{
Chung_2021()
}
\description{
Obtain the HeLa inCITE-seq data from Chung et al. (2021).
}
\details{
Chung \emph{et al} introduce a new method inCITE-seq (intranuclear CITE-seq)
enabling the measurement of intranuclear transcription factors. This data set
is a HeLa cell line in response to cytokine stimulation with a single
transcription factor (p65) profiled by DNA-conjugated antibodies.

The study also profiled a mouse brain sample in response to pharmacological
treatment.  The mouse samples are not included here.
}
