% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chung_inCITE.R
\name{Chung_inCITE}
\alias{Chung_inCITE}
\title{Obtain the Chung_inCITE data}
\usage{
Chung_inCITE()
}
\description{
Obtain the HeLa inCITE-seq data from Chung et al. (2021).
}
\details{
Chung \emph{et al} introduce a new method inCITE-seq (intranuclear CITE-seq)
enabling the measurement of intranuclear transcription factors. This data set
is a HeLa cell line in response to cytokine stimulation with a single
transcription factor (p65) profiled by DNA-conjugated antibodies.

The study also profiled a mouse brain sample in response to pharmacological
treatment.  The mouse samples are not included here.
}
\references{
Chung, Hattie, et al. "Joint single-cell measurements of nuclear proteins
and RNA in vivo." Nature methods 18.10 (2021): 1204-1212.
\doi{https://doi.org/10.1038/s41592-021-01278-1}
}
