% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Granja_2019.R
\name{Granja_2019}
\alias{Granja_2019}
\title{Obtain the Granja_2019 data}
\usage{
Granja_2019()
}
\description{
Obtain the CITE-seq data from Granja et al. (2019).
}
\details{
This study includes 16 CITE-seq samples of either PBMC,
BMMC (bone marrow mononuclear cells), CDC4 enriched BMMC, BM or MPAL
(mixed phenotype acute leukemia).  One of the MPAL patients was measured
before chemotherapy and 2 replicates after relapse.

Additional information the MPAL samples is available in Supplementary Table 2
of the original article, and antibody clone information is in
Supplementary Table 3.

Three isotype controls were listed in the clone table but are not in the data.
In addition, 17 ADTs are found in the data but not the clone table.

22 single cell ATAC-seq samples are also available but are not included here.

Citeseq sequenced across? 2 different sequencing platforms

#'CITE-seq data was downloaded from
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE139369}
}
\references{
Granja, Jeffrey M., et al. "Single-cell multiomic analysis
identifies regulatory programs in mixed-phenotype acute
leukemia." Nature biotechnology 37.12 (2019): 1458-1465.
\doi{10.1038/s41587-019-0332-7}
}
