% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hao_2021.R
\name{Hao_2021}
\alias{Hao_2021}
\title{Obtain the Hao_2021 data}
\usage{
Hao_2021()
}
\description{
Obtain the CITE-seq and ECCITE-seq data from Hao et al. (2021).
}
\references{
Hao, Yuhan, et al. "Integrated analysis of multimodal single-cell data."
Cell 184.13 (2021): 3573-3587.
\doi{https://doi.org/10.1016/j.cell.2021.04.048}
}
