% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hao_HIV.R
\name{Hao_HIV}
\alias{Hao_HIV}
\title{Obtain the Hao_HIV data}
\usage{
Hao_HIV(modality = c("both", "CITE", "ECCITE"))
}
\description{
Obtain the CITE-seq and ECCITE-seq data from Hao et al. (2021).
}
\details{
CITE-seq and ECCITE-seq of PBMCs from 8 volunteers enrolled in a HIV vaccine
trial.  A total of 211000 human PBMCs were collected from 8 volunteers
enrolled in a HIV vaccine trial.  Samples were collected immediately before
(Day 0), and 3 and 7 days after administration of vaccine.

For each experiment frozen PMBCs from four different patients at the 3
different time points were processed together.  Cells were then divided

The ECCITE-seq experiment used a combination of ECCITE-seq and TotalSeq C
reagents for CD45RA and CD45RO.

Information about the donors is available in the supplementary material
\url{https://ars.els-cdn.com/content/image/1-s2.0-S0092867421005833-mmc3.xlsx}
}
\references{
Hao, Yuhan, et al. "Integrated analysis of multimodal single-cell data."
Cell 184.13 (2021): 3573-3587.
\doi{https://doi.org/10.1016/j.cell.2021.04.048}
}
