% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Holmes_bcell.R
\name{Holmes_bcell}
\alias{Holmes_bcell}
\title{Obtain the Holmes_bcell data}
\usage{
Holmes_bcell()
}
\description{
Obtain the germinal centre B-cell CITE-seq data from Holmes \emph{et al.}
(2020).
}
\details{
CITE-seq of germinal centre B-cells isolated from tonsil tissue of a single
donor.  Germinal B-cells (CD3-/IgD-/CD38+) were sorted using FACS.

Counts for RNA and ADT reads were generated by CellRanger v2.1.0.

scRNA-seq was performed on B cells from two donors.  These samples are not
included in this data set.
}
\references{
Holmes, Antony B., et al. "Single-cell analysis of germinal-center B cells
informs on lymphoma cell of origin and outcome." Journal of Experimental
Medicine 217.10 (2020).
\doi{doi: 10.1084/jem.20200483}
}
