% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kotliarov.R
\name{Kotliarov}
\alias{Kotliarov}
\title{Obtain the Kotliarov data}
\usage{
Kotliarov()
}
\description{
Obtain the CITE-seq data from  \emph{et al.}
().
}
\details{
CD206 is present in the data but missing from the clone table.
}
\references{
\doi{}
}
