% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lawlor.R
\name{Lawlor}
\alias{Lawlor}
\title{Obtain the Lawlor data}
\usage{
Lawlor()
}
\description{
Obtain the CITE-seq data from Lawlor \emph{et al.}
().
}
\details{
CD185 is listed in the clone table but doesn't appear in the data
}
\references{
\doi{}
}
