% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Leader_2021.R
\name{Leader_2021}
\alias{Leader_2021}
\title{Obtain the Leader_2021 data}
\usage{
Leader_2021()
}
\description{
Obtain the lung cancer CITE-seq data from Leader et al. (2021).
}
\details{
Leader \emph{et al} used a mixture of commercially available CITE-seq
antibodies and in-house conjugated antibodies.

Additional RNA-seq and TCR-seq samples are available but not included in
this data set.
}
\references{
Leader, Andrew M., et al. "Single-cell analysis of human non-small cell
lung cancer lesions refines tumor classification and patient
stratification." Cancer Cell 39.12 (2021): 1594-1609.
\doi{10.1016/j.ccell.2021.10.009}
}
