% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mimitou_2021.R
\name{Mimitou_2021}
\alias{Mimitou_2021}
\title{Obtain the Mimitou_2021 data}
\usage{
Mimitou_2021()
}
\description{
Obtain the CITE-seq data from Mimitou et al. (2021).
}
\details{
Introduces a new method ASAP-seq (ATAC-seq with selected antigen profiling).
}
