% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Poch_liver.R
\name{Poch_liver}
\alias{Poch_liver}
\title{Obtain the Poch_liver data}
\usage{
Poch_liver()
}
\description{
Obtain the liver T-cell CITE-seq data from Poch \emph{et al.} (2021).
}
\details{
CITE-seq of intrahepatic T-cells AND BLOOD? from 6 patients with primary sclerosing
cholangitis (PSC), an immune-mediated liver disease.  T-cells were sorted from
liver tissue using FACs.

Count data were provided by the corresponding author. Fastq data is available
at
\url{https://www.ebi.ac.uk/biostudies/arrayexpress/studies?query=E-MTAB-10143}

scRNA-seq, scATAC-seq and scTCR-seq of additional samples is available
but not included in this data set.  The scRNA-seq includes liver and blood
samples from additional patients and b
}
\references{
Poch, Tobias, et al. "Single-cell atlas of hepatic T cells reveals expansion
of liver-resident naive-like CD4+ T cells in primary sclerosing cholangitis."
Journal of hepatology 75.2 (2021): 414-423.
\doi{10.1016/j.jhep.2021.03.016}
}
