% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PomboAntunes_glioblastoma_mm.R
\name{PomboAntunes_glioblastoma_mm}
\alias{PomboAntunes_glioblastoma_mm}
\title{Obtain the PomboAntunes_glioblastoma_mm data}
\usage{
PomboAntunes_glioblastoma_mm()
}
\description{
Obtain the mouse glioblastoma single cell CITE-seq data from
Pombo Antunes et al (2021).
}
\details{
CITE-seq of CD45+ cells from GL261 tumors from wildtype and
Ccr2-knockout mice (tumors from three individual mice pooled per sample).

Information on mice is available in the Research Reporting Summary of
the original paper and the mouse CITE-seq antibody panel is in
Supplementary Table 14.

Additional data available but not included here:
\itemize{
\item{scRNA-seq for 7 patients with newly diagnosed and 4 patients with
recurrent glioblastoma}
\item{scRNA-seq for CD45+ immune cells from GL261 tumors (mouse).}
}
Data was downloaded from
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE163120}
#'
}
\references{
Pombo Antunes, Ana Rita, et al. "Single-cell profiling of myeloid cells
in glioblastoma across species and disease stage reveals macrophage
competition and specialization." Nature neuroscience 24.4 (2021): 595-610.
\doi{https://doi.org/10.1038/s41593-020-00789-y}
}
\seealso{
Code:
\url{https://github.com/Movahedilab/Glioblastoma}
Interactive data explorer:
\url{www.brainimmuneatlas.org}
}
