% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Buus_2021.R
\name{Buus_2021}
\alias{Buus_2021}
\title{Obtain the Buus_2021 data}
\usage{
Buus_2021()
}
\description{
Obtain the single cell CITE-seq data from healthy donors and a lung tumor
patient from Buus et al (2021).
}
\details{
CITE-seq of PBMCs from healthy donors and lung tumour cells from a
single patient.  Different antibody staining conditions were tested.

Data was downloaded from
\url{https://doi.org/10.6084/m9.figshare.c.5018987}
}
\references{
Buus, Terkild B., et al. 'Improving oligo-conjugated
antibody signal in multimodal single-cell analysis.'
Elife 10 (2021): e61973.
}
\seealso{
\url{https://github.com/Terkild/CITE-seq_optimization}
}
